P4C-1484: rewrite match power config based on predecessor tables (#4120)
- refactor/simplify mpr data structures
- fix warnings
- Reduce use of mpr_bus where not needed
- avoid unnecessary conflicts when theads differ as to match/action
  dependency
- don't auto-promote everything to match dependent; only if there is an
  actual conflict in the mpr_glob_exec or long_branch busses.
- Fix long branch programming in the assembler:  tables may drive
  mulitple long branches, so look for sets that give
  maximal coverage of the tables we're trying to enable, avoiding
  last-resort usage of next table pointers (which we don't check for
  conflicts)
- new tests

+++ bf-p4c/mau/jbay_next_table.cpp
@@ -529,7 +529,7 @@ bool JbayNextTable::Prop::preorder(const IR::MAU::Table* t) {
 }
 
 void JbayNextTable::Prop::end_apply() {
-    for (unsigned i = 0; static_cast<int>(i) < self.max_stage; ++i) {  // Fix up stage_id
+    for (int i = 0; i < self.max_stage; ++i) {  // Fix up stage_id
         if (self.stage_id[i] < i * Memories::LOGICAL_TABLES)
             self.stage_id[i] = i * Memories::LOGICAL_TABLES;
         else
@@ -790,7 +790,7 @@ struct JbayNextTable::TagReduce::merge_t {
 
 JbayNextTable::TagReduce::merge_t JbayNextTable::TagReduce::merge(Tag l, Tag r) const {
     merge_t rv;
-    std::map<int, unsigned> stage_id(self.stage_id);  // Need our own copy
+    std::map<int, int> stage_id(self.stage_id);  // Need our own copy
     auto cap = [&](int k) {  // Check the capacity of a stage. 0 means full
                    return (k+1) * Memories::LOGICAL_TABLES - stage_id[k];
                };
